==> 05 Land use extractions.Rmd <==
---
title: "Land use extractions"
author: "Alice Armstrong"
date: "2023-03-09"
output: html_document

This script takes extracts various types of land use data from the error polygons of the specimens.


---
#Load packages#
```{r}
library(sfe)
library(sf)
library(rgeos)
library(rgdal)
library(Hmisc)
library(tidyverse)
library(here)
library(patchwork)
library(knitr)
library(ncdf4)
library(patchwork)
library(raster)
```

#Increase the R memory limit for big files#
```{r}
gc()
memory.limit(size=500000)
```

#Load custom function#
#Make sure your custom function is saved as an ‘R’ script not an ‘Rmarkdown’ script#
```{r}
source(here::here("04 Functions anthropocene data.R"))
```

#Read in the georeferenced data#
```{r}
crocs<-read.csv('full_specimin_localility')
```

#Convert to sf and rename extent column on full data set#
```{r}
crocs <- 
  crocs %>% rename('Extent_km'=Extent)%>%
  rename(ID = 'X')%>%
  st_as_sf(coords = c("Long", "Lat"), crs = 4326)
```

##------------------------------------------------
## Add error polygons to data
##------------------------------------------------
# add a `buffer` to the point data using the point-radius method.
# Ensure it is in the correct coord system for Africa to get distances correct
```{r}
crocs_errors <- st_transform(crocs, "+proj=chamb +lat_1=22 +lon_1=0 +lat_2=22 +lon_2=45 +lat_3=-22 +lon_3=22.5 +datum=WGS84 +type=crs + units=km")
crocs_errors <- st_buffer(crocs, crocs$Extent_km)
```

# Reconvert to CRS = 4326 for mapping#
```{r}
crocs_errors <- st_transform(crocs_errors, crs = 4326)
```

#------#
```{r}
crocs_errors <- st_transform(crocs, "+proj=chamb +lat_1=22 +lon_1=0 +lat_2=22 +lon_2=45 +lat_3=-22 +lon_3=22.5 +datum=WGS84 +type=crs + units=km")
crocs_errors <- st_buffer(crocs, crocs$Extent_km)
```

#Reconvert to CRS = 4326 for mapping#
```{r}
crocs_errors <- st_transform(crocs_errors, crs = 4326)
```

## Save the outputs
# Note that `st_write` will throw an error if you try to overwrite a file, 
# so delete previous version of the files first. 
# To avoid this error I've commented these out here.

```{r}
st_write(crocs, "crocs-points.csv", layer_options = "GEOMETRY=AS_WKT")
st_write(crocs_errors, "crocs-extents.csv",layer_options = "GEOMETRY=AS_WKT")

```


#Extracting land use data#
#Read in the netCDF data#
```{r}
nc_data <- nc_open(here("C:/Users/alice/Downloads/states.nc"))
```

#Capture the lat, lon and time variables#
```{r}
lon <- ncvar_get(nc_data, "lon")
lat <- ncvar_get(nc_data, "lat", verbose = F)
time <- ncvar_get(nc_data, "time")

overlaps_landuse<-crocs_errors
```

#Primary forest#
```{r}
overlaps_landuse <- 
  get_landuse_diff("primf", nc_data, overlaps_landuse) 
```

#Primary non forest#
```{r}
overlaps_landuse <- 
  get_landuse_diff("primn", nc_data, overlaps_landuse) 
```

#Urban#
```{r}

overlaps_landuse <- 
  get_landuse_diff("urban", nc_data, overlaps_landuse) 
```

#Close data set#
```{r}
nc_close(nc_data)
```

# Write to file#
```{r}
st_write(overlaps_landuse, here("data/overlaps_landuse.csv"))

```
